rename registration to cloud_registration
