+++ lidar_mapping/include/lidar_mapping/loop_closure.hpp
@@ -22,7 +22,7 @@
 #include <string>
 //
 #include "localization_common/cloud_filter/cloud_filter_factory.hpp"
-#include "localization_common/registration/registration_factory.hpp"
+#include "localization_common/cloud_registration/cloud_registration_factory.hpp"
 #include "localization_common/sensor_data/key_frame.hpp"
 #include "localization_common/sensor_data/loop_pose.hpp"
 #include "scan_context/scan_context_manager.hpp"
@@ -66,9 +66,9 @@ private:
 
   std::shared_ptr<localization_common::CloudFilterInterface> current_scan_filter_;
   std::shared_ptr<localization_common::CloudFilterInterface> map_filter_;
-  std::shared_ptr<localization_common::RegistrationInterface> registration_;
+  std::shared_ptr<localization_common::CloudRegistrationInterface> registration_;
   //
-  std::shared_ptr<localization_common::RegistrationFactory> registration_factory_;
+  std::shared_ptr<localization_common::CloudRegistrationFactory> registration_factory_;
   std::shared_ptr<localization_common::CloudFilterFactory> cloud_filter_factory_;
   //
   std::shared_ptr<scan_context::ScanContextManager> scan_context_manager_;
