+++ lidar_mapping/src/loop_closure.cpp
@@ -26,7 +26,7 @@ namespace lidar_mapping
 {
 LoopClosure::LoopClosure()
 {
-  registration_factory_ = std::make_shared<localization_common::RegistrationFactory>();
+  registration_factory_ = std::make_shared<localization_common::CloudRegistrationFactory>();
   cloud_filter_factory_ = std::make_shared<localization_common::CloudFilterFactory>();
 }
