+++ lidar_odometry/src/front_end.cpp
@@ -27,7 +27,7 @@ FrontEnd::FrontEnd()
 : local_map_(new localization_common::PointXYZCloud()),
   result_cloud_(new localization_common::PointXYZCloud())
 {
-  registration_factory_ = std::make_shared<localization_common::RegistrationFactory>();
+  registration_factory_ = std::make_shared<localization_common::CloudRegistrationFactory>();
   cloud_filter_factory_ = std::make_shared<localization_common::CloudFilterFactory>();
 }
