+++ localization_common/src/cloud_registration/icp_registration.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "localization_common/registration/icp_registration.hpp"
+#include "localization_common/cloud_registration/icp_registration.hpp"
 
 namespace localization_common
 {
