+++ localization_common/src/cloud_registration/ndt_omp_registration.cpp
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "localization_common/registration/ndt_registration.hpp"
+#include "localization_common/cloud_registration/ndt_omp_registration.hpp"
 
 namespace localization_common
 {
 
-NDTRegistration::NDTRegistration(const YAML::Node & node)
-: ndt_(new pcl::NormalDistributionsTransform<PointXYZ, PointXYZ>())
+NDTOmpRegistration::NDTOmpRegistration(const YAML::Node & node)
+: ndt_(new pclomp::NormalDistributionsTransform<PointXYZ, PointXYZ>())
 {
   float res = node["res"].as<float>();
   float step_size = node["step_size"].as<float>();
@@ -28,20 +28,20 @@ NDTRegistration::NDTRegistration(const YAML::Node & node)
   set_param(res, step_size, trans_eps, max_iter);
 }
 
-NDTRegistration::NDTRegistration(float res, float step_size, float trans_eps, int max_iter)
-: ndt_(new pcl::NormalDistributionsTransform<PointXYZ, PointXYZ>())
+NDTOmpRegistration::NDTOmpRegistration(float res, float step_size, float trans_eps, int max_iter)
+: ndt_(new pclomp::NormalDistributionsTransform<PointXYZ, PointXYZ>())
 {
   set_param(res, step_size, trans_eps, max_iter);
 }
 
-bool NDTRegistration::set_param(float res, float step_size, float trans_eps, int max_iter)
+bool NDTOmpRegistration::set_param(float res, float step_size, float trans_eps, int max_iter)
 {
   ndt_->setResolution(res);
   ndt_->setStepSize(step_size);
   ndt_->setTransformationEpsilon(trans_eps);
   ndt_->setMaximumIterations(max_iter);
-
-  std::cout << "NDT params:" << std::endl
+  ndt_->setNumThreads(8);
+  std::cout << "NDT OMP params:" << std::endl
             << "res: " << res << ", "
             << "step_size: " << step_size << ", "
             << "trans_eps: " << trans_eps << ", "
@@ -51,14 +51,14 @@ bool NDTRegistration::set_param(float res, float step_size, float trans_eps, int
   return true;
 }
 
-bool NDTRegistration::set_input_target(const PointXYZCloudPtr & input_target)
+bool NDTOmpRegistration::set_input_target(const PointXYZCloudPtr & input_target)
 {
   ndt_->setInputTarget(input_target);
 
   return true;
 }
 
-bool NDTRegistration::match(
+bool NDTOmpRegistration::match(
   const PointXYZCloudPtr & input_source, const Eigen::Matrix4f & predict_pose,
   PointXYZCloudPtr & result_cloud, Eigen::Matrix4f & result_pose)
 {
